Feat: add admin && Refactor

+++ src/main/java/com/luna/TodoList/service/UserService.java
@@ -1,5 +1,6 @@
 package com.luna.TodoList.service;
 
+import com.luna.TodoList.dto.UserPublicDto;
 import com.luna.TodoList.dto.UserRequestDto;
 import com.luna.TodoList.exception.NotFoundException;
 import com.luna.TodoList.model.User;
@@ -7,50 +8,68 @@ import com.luna.TodoList.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
-    private static UserRepository userRepository;
+    private final UserRepository userRepository;
 
     public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
 
-    public User addUser(UserRequestDto userRequestDto) {
-        User user = User.builder()
-                .username(userRequestDto.getUsername())
-                .DateOfBirth(userRequestDto.getDateOfBirth())
-                .email(userRequestDto.getEmail())
-                .phone(userRequestDto.getPhone())
-                .build();
-        userRepository.save(user);
-        return user;
+    public String deleteUserById(Long id, UserRequestDto userRequestDto) throws Exception {
+
+        Long userLogin = userRepository.findByUsername(userRequestDto.getUsername()).getUserId();
+
+        if (id.equals(userLogin) || userRequestDto.getAdmin()) {
+            userRepository.findById(id).orElseThrow(() -> new NotFoundException("User not exist"));
+            userRepository.deleteById(id);
+            return "SUCCESS";
+        } else {
+            throw new Exception("You do not have access");
+        }
     }
 
-    public String deleteUserById(Long id) throws NotFoundException{
-        userRepository.findById(id).orElseThrow(() -> new NotFoundException("User not exist"));
-        userRepository.deleteById(id);
-        return "SUCCESS";
+    public User updateUser(Long id, UserRequestDto userRequestDto) throws Exception {
+
+        Long userLogin = userRepository.findByUsername(userRequestDto.getUsername()).getUserId();
+
+        if (id.equals(userLogin) || userRequestDto.getAdmin()) {
+            userRepository.findById(id).orElseThrow(()-> new NotFoundException("User is not exist!"));
+            User userToUpdate = userRepository.getOne(id);
+            userToUpdate.setUsername(userRequestDto.getUsername());
+            userToUpdate.setDateOfBirth(userRequestDto.getDateOfBirth());
+            userToUpdate.setEmail(userRequestDto.getEmail());
+            userToUpdate.setPhone(userRequestDto.getPhone());
+            userRepository.save(userToUpdate);
+            return userToUpdate;
+        } else {
+            throw new Exception("You do not have access");
+        }
     }
 
-    public User updateUser(Long id, UserRequestDto userRequestDto){
-        userRepository.findById(id).orElseThrow(()-> new NotFoundException("User is not exist!"));
-        User userToUpdate = userRepository.getOne(id);
-        userToUpdate.setUsername(userRequestDto.getUsername());
-        userToUpdate.setDateOfBirth(userRequestDto.getDateOfBirth());
-        userToUpdate.setEmail(userRequestDto.getEmail());
-        userToUpdate.setPhone(userRequestDto.getPhone());
-        userRepository.save(userToUpdate);
-        return userToUpdate;
+    public List<User> getAllUsers(Long loginId) throws Exception {
+        if (userRepository.findAdminByUserId(loginId)) {
+            return userRepository.findAll();
+        } else {
+            throw new Exception("You do not have access");
+
+        }
     }
 
-    public List<User> getAllUsers(){
-        return userRepository.findAll();
+    public User getUserById(Long id, UserRequestDto userRequestDto) throws Exception {
+
+        Long userLogin = userRepository.findByUsername(userRequestDto.getUsername()).getUserId();
+
+        if (id.equals(userLogin) || userRequestDto.getAdmin()) {
+            return userRepository.findById(id).orElseThrow(()-> new NotFoundException("User not exit"));
+        } else {
+            throw new Exception("You do not have access");
+        }
     }
 
-    public User getUserById(Long id){
-        return userRepository.findById(id).orElseThrow(()-> new NotFoundException("User not exit"));
+    public UserPublicDto getOtherUserById(Long id, Long usersInfoId) throws Exception {
+        return UserPublicDto.builder().username(userRepository.findUsernameById(usersInfoId)).build();
     }
-
-
 }
